afs: avoid GetEntryByName with .readonly extension
Currently, the volume name hash table is indexed by the volume name
without any extension (*.readonly / *.backup). As a result, the function
responsible for finding the vl entry for the given volume (FindByName)
removes these extensions from the volume name. This means that, for a
volume called xxx, FindByName(xxx) and FindByName(xxx.readonly) will
return the same record. That said, if afs_GetVolumeByName(xxx) failed
because xxx could not be found in the vldb,
afs_GetVolumeByName(xxx.readonly) will have the same end result.

Therefore, this extra call is not necessary and can be removed. This
change can slightly improve the response time of the client if the
volume in question does not exist.