ubik: don't set db epoch to zero in SDISK_SendFile
According to commit 161a1f65095eb6da989a7386588918c9ec5ef640, failures
in SDISK_SendFile can potentially leave remote sites with a broken
database. Knowing that corrupted data should not be served to the
clients, the commit mentioned above sets the database epoch to zero
before receiving any data from the sync-site. As a result, if
SDISK_SendFile fails, the database epoch is left as zero and subsequent
transactions will not be allowed on the remote in question.

The idea behind this patch was correct until the introduction of commit
0f9529171c37c4ef1e76156da111d83d04b38505. With this change, the data
received from the sync-site is not written directly into the database.
Instead, a temporary file is used and the live database will only be
replaced by this file if SDISK_SendFile succeeds.

With that being said, failures in SDISK_SendFile do not result in a
truncated / corrupted database anymore and the remote site in question
should be able to serve read transactions even if SDISK_SendFile does
not succeed. To make it possible, do not set the database epoch to zero
at the beginning of this RPC.