ubik: invalidate database if rename() failed
Currently, the synchronization-site sends a copy of its database to the
remote-sites through SDISK_SendFile. This RPC writes the data received
into a temporary file which will only replace the live database if all
the data sent by the synchronization-site were successfully received. In
order to achieve the replacement mentioned above, the system call
rename() is used to change the name of the temporary file to the name
used by live databases. Unfortunately, failures on this system call are
not handled correctly. If rename() fails, we do not have any guarantee
that we have a valid database. As a result, the remote-site in question
should not serve any data to the clients. To fix this problem,
invalidate database if rename() failed.