ubik: abort invalid transactions in advance
Currently, transactions from previous mandates are aborted by
urecovery_CheckTid. However, we do not have to wait for this function to
abort invalid transactions. If SDISK_GetFile is being executed, we know
that a new quorum was just elected and, at this point, write
transactions should not exist. As a result, if we have a write
transaction on the remote-site that is running SDISK_GetFile, we know
that this transaction is from the last mandate and should be aborted.

Instead of waiting for urecovery_CheckTid, let SDISK_GetFile abort
transactions from previous mandates in advance.