ubik: do not unnecessarily invalidate database
According to commit 161a1f65095eb6da989a7386588918c9ec5ef640, failures
in SDISK_SendFile can potentially leave the remote site in question with
a truncated / corrupted database. As a result, this site should not keep
serving data after failures in this RPC. To make it possible, the commit
mentioned above introduced a patch that sets the database epoch to zero
before we receive any data from the sync-site. If SDISK_SendFile does
not succeed, the epoch is left as zero and subsequent transactions will
not be allowed on this site.

The idea behind this patch was correct until the introduction of commit
0f9529171c37c4ef1e76156da111d83d04b38505. With this change, the data
received from the sync-site is not written directly into the database.
Instead, a temporary file is used and the live database will only be
replaced by this file if SDISK_SendFile succeeds. Similarly, the
recovery thread does not touch its local database while the sync-site
is receiving a new copy from another peer.

With that being said, failures during this transfer do not result in a
truncated / corrupted database anymore and the site in question should
be able to serve read transactions even if the reception of the new
database does not succeed. To make it possible, do not invalidate our
local copy before starting the transfer of the new database.

Change-Id: Iaf9cf8c5cc95e94923c76250802ab30cd14c89ee