LINUX: Honor --enable-checking for libafs
When we build the kernel module on LINUX, we don't pass in any of our
CFLAGS, since the Linux buildsystem itself figures out what flags are
needed. However, this means that we don't pass in -Werror when
--enable-checking is turned on, so warnings may not cause the build to
fail.

To fix this, create a new autoconf variable, called CFLAGS_WERROR,
that only contains -Werror if --enable-checking is turned on. We then
pass that into the Linux module buildsystem, so -Werror is given to
the compiler when building our module.

Reviewed-on: https://gerrit.openafs.org/13682
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
(cherry picked from commit 6e0f1c3b45102e7644d25cf34395ca980414317f)

[cwills@sinenomine.net] On master gerrit 14106 comes after this commit.
In the 1.8.x branch the 14106 commit is already pulled in (gerrit 14217)
which causes a conflict due a difference ('-fno-common') in the line
above the change in osconf.m4.

Change-Id: I4e49d5c5fdf26399eb04d2f76196f3c3f4e7baf6