volser: remsite must give ro id to VL_DeleteEntry
Currently, vos remsite can be used to remove a read-only site definition
from a vldb entry. If this site definition is the last reference to a
volume (read-write site definition is not present), the whole vldb entry
is removed. In this case, ubik_VL_DeleteEntry(), which receives the type
and volume id as arguments, is called. If the volume type argument is
ROVOL, then a read-only volume id is expected (otherwise the vldb entry
will not be found and this function will fail).

In the scenario mentioned above, the current version of remsite tries to
remove the vldb entry associated with a read-only volume (ROVOL) calling
ubik_VL_DeleteEntry(). Unfortunately, the volume id given to this rpc
is a read-write volume id. As a result, ubik_VL_DeleteEntry() fails and
vos remsite returns (not removing the entry and leaving the vldb entry
in question locked).

To fix this problem, provide the correct arguments (read-only volume id)
to ubik_VL_DeleteEntry().

Change-Id: I7b3d1d0d455e0af4001df6240c85005bc7a3a3f3