vol: allow attachment of partitions dynamically
Currently, it is not possible to add new partitions to running servers
dynamically. In other words, new vice partitions added to the system
after the initialization of the fileserver are completely ignored. As a
workaround, sysadmins can 'force' the attachment of the new partitions
restarting the fileserver. Unfortunately, rebooting the live servers is
not always an option.

This commit introduces a new feature that allows the attachment of new
vice partitions dynamically. To make it possible, a new 'vos' command
was created. This command forces the fileserver, volume server, and
salvager server to scan and attach new mounted vice partitions.

Change-Id: I64b72f1868fdc651c23f2e66ea6561adf4e665c0