Remove unused rxkad_GetServerInfo parameters
A few callers were giving pointers to rxkad_GetServerInfo to extract
information the caller didn't actually need (that is, for the security
level, expiration, or kvno for the conn). Change these callers to pass
NULL for the relevant parameters, to make it clearer in the code that
this info is not used.

Change-Id: I7b44a1f994fb2f297a3bfe857ac24b2359b708fe