rx: Introduce rx_GetConnSecId and friends
Currently, there are many areas in the code that do something like the
following pseudocode to get security-related information from a
connection for a running call:

    secClass = rx_SecurityClassOf(conn);
    if (secClass == RX_SECIDX_KAD) {
        code = rxkad_GetServerInfo(conn, &level, &expTime, name, inst, cell, &kvno);
        code = afsconf_IsLocalRealmMatch(dir, &islocal, name, inst, cell);
        if (islocal) {
            fullname := "name.inst";
        } else {
            fullname := "name.inst@cell";
        }
    }

That is, every call site that wants to get this information (security
level, expiration time, caller principal name) must check the security
index, and call a security-class-specific function. And then, many
call sites must also separately call afsconf_IsLocalRealmMatch to see
if the caller is a "local" user or not, and we construct the full name
of the user.

Doing all of this logic at every call site is very cumbersome, and
makes it difficult to adjust all of these call sites for different
security classes.

To improve this, introduce several security-class-agnostic functions,
to extract these different pieces of information: rx_GetConnSecLevel,
rx_GetConnSecExpiration, and rx_GetConnSecId. Implement these for the
rxkad security class. Also introduce a new afsconf function,
afsconf_Krb4LocalIdentity, which performs the "local realm" check and
related transformation for krb4-based identities.

Adjust most callers to use these new functions. The new functions for
getting the caller identity no longer split the caller information
between name, instance, and cell/realm, so a few callers need some new
additional logic to extract the cell/realm (none of the adjusted sites
care about the instance being split out from the name).

Do not adjust the following callers of rxkad_GetServerInfo:

- src/budb/procs.c
- src/kauth/kaprocs.c
- src/update/server.c

These sites require some additional effort to work with the new
functions, but the relevant daemons are rarely used, so this commit
does not handle them. They can always be handled later on in a
different commit, if support for additional security classes in those
areas is desired. Alternatively, if the code for those callers
eventually goes away, then rxkad_GetServerInfo itself can be removed.

Change-Id: I5dcb62dcd526fc5ce3b19fe2a99b5b2ef2298440