volser: remsite must give rw id to VL_DeleteEntry
Currently, vos remsite can be used to remove a read-only site definition
from a vldb entry. If this site definition is the last reference to a
volume (read-write site definition is not present), the whole vldb entry
is removed. In this case, ubik_VL_DeleteEntry(), which receives the type
and volume id as arguments, is called. If the volume type argument does
not match the type of the volume identified by the volume id, the vldb
entry in question will not be found and this function will return an
error.

In the scenario mentioned above, the current version of remsite tries to
remove the vldb entry in question calling ubik_VL_DeleteEntry().
Unfortunately, the volume id given to this rpc is a read-write volume id
but the volume type, also provided as an argument, is ROVOL. As a
result, ubik_VL_DeleteEntry() does not succeed.

To fix this problem, provide the correct arguments (read-write volume
type) to ubik_VL_DeleteEntry(). Notice that vldb entries are mainly
identified by read-write volume ids (not read-only).

Change-Id: I7b3d1d0d455e0af4001df6240c85005bc7a3a3f3