afs: Add rxgk support for RXAFS communication
[XXX: Note the FIXME in rxgk_packet.c.]

Update afs_pickSecurityObject to make use of rxgk tokens, if the
calling user has them. Doing this is slightly complicated by the fact
that negotiating fileserver rxgk tokens itself needs to create a new
rxgk-authenticated connection to the cell's vlservers, and so needs to
call back into the afs_Conn* subsystem.

So in order to do this, adjust the semantics of afs_pickSecurityObject
a bit:

- Allow afs_pickSecurityObject to return an actual error code.

- Add a new output argument, called 'need_net'. If the caller gives a
  non-NULL pointer, it indicates we are holding afs_xconn (and
  afs_pickSecurityObject cannot call afs_Conn*). If
  afs_pickSecurityObject sets *need_net on return, it indicates that
  we need to call back into afs_Conn* to negotiate the security
  object, and so the caller can call afs_pickSecurityObject again with
  afs_xconn dropped (and a NULL 'need_net').

Currently, this commit falls back to other security classes (rxkad, or
rxnull), if we encounter an error during rxgk token negotiation. The
rxgk spec specifically says not to do this, in order to avoid
downgrade attacks, but we allow downgrades for now (and log a warning)
while rxgk support is still experimental.

Change-Id: I64d74bfa46ab09ff288b3d7fdb2e35428f271813