rxgk: Convert RXGK_ autoconf vars to RXGK_ONLY
Currently, we have a few autoconf vars that are set conditionally,
depending on whether or not rxgk is enabled (such as RXGK_LIBS,
RXGK_LIBS_RPC, etc). This works, but defining the values for these
vars is done very far away from where they are used, and this means
that for any part of a Makefile that is conditional on rxgk being
enabled, we must define a new autoconf var in src/cf/gss.m4.

To shift these definitions closer to where they are used, and arguably
simplify the process, convert these variables to use an
@RXGK_ONLY@-style approach. That is, in our autoconf, we AC_SUBST a
variable called RXGK_ONLY. When rxgk is not enabled, this gets defined
to '#' (that is, a single octothorpe), and then rxgk is enabled, it
gets defined to an empty string. This way, a Makefile written like
this:

    @RXGK_ONLY@LIBS=$(top_builddir)/src/rxgk/liboafs_rxgk.la

will expand to this when rxgk is disabled:

    #LIBS=$(top_builddir)/src/rxgk/liboafs_rxgk.la

and will expand to this when enabled:

    LIBS=$(top_builddir)/src/rxgk/liboafs_rxgk.la

This allows us to use rxgk-related libraries and cflags closer to how
we'd normally use library libs and cflags (just with some lines
prefixed with @RXGK_ONLY@).

Provide a bit of documentation for this new subst var in the
OPENAFS_GSS autoconf function. Also document some of the other
rxgk-related autoconf bits while we're here to explain what all of the
vars mean.

Change-Id: I02db79473395d7e10281d5e5e9f8151121f999c8