volscan: cache dirent names to improve performance
Volscan produces output for each type of vnode object found in a volume.
By default, this tool provides the path of each vnode relative to the
volume root directory. To do so, volscan performs an inverse lookup on
the parent directory of each vnode to find the dirent string of the
vnode in question. As an example, if a given directory contains
references to 'n' files, 'n' inverse lookups will be performed on this
directory. Unfortunately, this approach results in many accesses to the
disk, affecting the response time of this tool.

To improve this situation, avoid redundant inverse lookups by caching
dirent strings stored by each directory.

Change-Id: Ib1d63c9ef1ef0f677b1df15e0bb637e7d86f12a4