ubik: fix SVOTE_GetSyncSite output argument
The SVOTE_GetSyncSite RPC provides the IP address of the current
sync-site. This address is stored into ahost, one of the output
arguments of this remote procedure call. Unfortunately, ubik_int.xg
defines ahost as an input argument. As a result, the IP address in
question is not returned to the callers of SVOTE_GetSyncSite.

To fix this problem, define ahost as an output argument.

Change-Id: I2be9d90cf33302fc235857da02f073eb82f23d20