ubik: don't resend db if site didn't vote for us
Currently, the recovery procedure is triggered in the following
situations:

1a. immediately after an election;
2a. when a remote site comes up;
3a. when the db of a remote site is not up to date;

During this procedure, the coordinator pools all servers (in quorum) to
find which one holds the best database version. After retrieving this
version (if necessary), the sync-site propagates this database to the
other sites. Unfortunately, this expensive procedure is triggered
repeatedly in some specific situations. Consider the following scenario:

1b.  a remote site holding an outdated db comes up;
2b.  this site won't vote for any site in the next BIGTIME seconds (ubik
     time constraints);
3b.  the recovery procedure is executed and the current db is sent to
     this site;
4b.  incoming write tx isn't propagated to this site (it didn't vote for
     the coordinator yet);
5b.  the remote in question is set as outdated (missed write tx);
6b.  goto 3b;
...
7b.  after BIGTIME seconds, this remote site votes for the coordinator;
8b.  the recovery procedure is executed and the current db is sent to
     this site;
9b.  at this point, incoming write txs are propagated to this site;
10b. done;

In the scenario mentioned above (common in cells with frequent write
transactions), the expensive recovery procedure is executed several
times in the BIGTIME seconds timeframe. In order to avoid several
resyncs during this period, execute recovery procedure to update
outdated sites only if the receiver voted for the coordinator. With this
change, the scenario mentioned above would result in the following
behavior:

1c.  a remote site holding an outdated db comes up;
2c.  this site won't vote for any site in the next BIGTIME seconds (ubik
     time constraints);
3c.  the recovery procedure is executed and the current db is sent to
     this site (2a);
4c.  incoming write tx isn't propagated to this site (it didn't vote for
     the coordinator yet);
5c.  the remote in question is set as outdated (missed write tx);
6c.  since this site didn't vote for the coordinator yet, the recovery
     procedure isn't triggered;
7c.  after BIGTIME seconds, this remote site votes for the coordinator;
8c.  the recovery procedure is executed and the current db is sent to
     this site;
9c.  at this point, incoming write txs are propagated to this site;
10c. done;

Therefore, this new behavior avoids several expensive recoveries.
Instead of resending the database repeatedly during the BIGTIME seconds
timeframe, the resync operation is performed only twice.

Note: The behavior described in 4b and 4c was introduce by commit
68ec78950a6e39dc1bf15012d4b889728086d0b7. Without this change, write
transactions are also propagated to sites that didn't vote for the
coordinator. However, since the receiver didn't vote for the requestor,
the transaction is refused, resulting in the same behavior as the one
described in this commit message.

Change-Id: I9599130b1545b6fadc22d456a4df7ca8a2b085a0