vos: add new option to escalate listvldb request
Currently, ubik read requests can be served by any valid version of the
ubik database (even if this version is not current). As a result, reads
served by outdated versions may not reflect the current state of the
database in question. In most cases, the outdated version differs from
the current db by only a few transactions. As a result, the majority of
the read requests can be successfully served by any ubik database that
holds a valid version. However, the few scenarios where the outdated db
does not have an entry that exists may induce users / automated scripts
to inappropriate behaviors.

To avoid the problem mentioned above, introduce a new option
(-escalate-search) that propagates vos reads to other vlservers if the
volume provided as an argument could not be found by the server that
first received this request. To limit the scope of these changes, only
listvldb can take advantage of this feature. Other vos operations may
offer this option in the future.

Change-Id: I1dfb1074fb74339c8c911f9d525d52a69b2801d2