ubik: death to SVOTE_GetSyncSite
The SVOTE_GetSyncSite RPC provides the IP address of the current
sync-site. This address is stored into ahost, one of the output
arguments of this remote procedure call. Unfortunately, ubik_int.xg
defines ahost as an input argument. As a result, the IP address in
question is not returned to the callers of SVOTE_GetSyncSite.

Since this argument can't be fixed without causing compatibility issues
between different client / server versions, remove this remote procedure
call.

Change-Id: Idd071583e8f67109e003f7a5675de02a235e5809