vol: allow attachment of partitions dynamically
Currently, it is not possible to add new partitions to running servers
dynamically. In other words, new vice partitions added to the system
after the initialization of the fileserver are completely ignored. The
only way to add new partitions is to restart the fileserver process, but
an administrator may not want to restart a fileserver that is actively
being accessed.

This commit introduces a new feature that allows the attachment of new
vice partitions dynamically. To make it possible, a new 'vos' command
was created. This command forces the fileserver, volserver, and
salvageserver to scan and attach new mounted vice partitions.

In order to implement this, this commit makes it so traversing the
DiskPartitionList requires VOL_LOCK.

Change-Id: I64b72f1868fdc651c23f2e66ea6561adf4e665c0