vos: add new option to send reads to the sync-site
Currently, ubik read requests can be served by any valid version of the
ubik database (even if this version is not current). As a result, reads
served by outdated versions may not reflect the current state of the
database in question. In most cases, the outdated version differs from
the current db by only a few transactions. As a result, the majority of
the read requests can be successfully served by any ubik database that
holds a valid version. However, the few scenarios where the outdated db
does not have an entry that exists may induce users / automated scripts
to inappropriate behaviors.

To avoid the problem mentioned above, introduce a new option (-sync)
that forces vos to send its ubik read requests to the sync-site. Doing
so, we increase our chances of being served by the most recent version
of the ubik database.

In order to preserve the benefits of load balancing, this option must be
reserved for important administrative operations.

Change-Id: I3a6022b759e20c3242df1c3c4a54d5b1344c96fc