vol-info: store vnode index info in globals
Currently, the vol-info library provides a set of functions to extract
data from volumes. To get information about the vnodes of a specific
volume, the index file of each type of vnode (large and small) is opened
and closed as needed. However, if the path of each vnode is requested
(NeedDirIndex), a copy of the large vnode index handle is kept in a
global (DirIndexFd) to be used latter (GetDirVnode).

To be consistent and keep the code simpler, store vnode index info in
globals. Also, move code that processes vnodes from HandleVnodes() to a
new function.

Change-Id: I9dbf0b34237d7508c33bec3cf0aa1d7ee915e790