afs: rework -linkedcell feature
Currently, if the -linkedcell feature is on, the afs client redirects to
the linked cell requests for volumes that couldn't be found in our local
cell. Unfortunately, requests that can’t reach our local vlserver won’t
be redirected. In other words, if our local vlserver is down, we don’t
know if we have the requested volume locally and we don’t redirect this
request to the linked cell. To work around this limitation, try to find
the requested volume in the linked cell if our local servers can’t be
reached.

Change-Id: I1c1a1c56b212f37d695aabaa171ad13308ded8f6