afs: pass copy of fid to afs_Conn and afs_Analyze
Currently, the communication from the afs client to the fileserver works
as follows:

1. A connection is created (afs_Conn);
2. Remote procedure call is executed (RXAFS_*);
3. The outcome of this RPC is analyzed (afs_Analyze);
4. Depending on the result returned by step 3, come back to step 1;

In preparation for a future commit, pass a local copy of the fid to the
functions mentioned above. Doing so, the values of this fid (volume id)
can be modified by afs_Conn, depending on the outcome of the RPC called
in step 2.

No functional change is incurred by this commit.

Change-Id: Ic144eeae06a03b8129382cf53dadfad41ce66fdf