afs: pass cell pointer address to afs_Analyze
Currently, the last argument of afs_Analyze is a pointer to a cell
struct. In preparation for a future commit, pass the address of this
pointer to this function. Doing so, afs_Analyze can modify this argument
(if necessary) depending on the outcome of the RPC in question.

No functional change is incurred by this commit.

Change-Id: I500a8c6db17bf2d602e7ec226a7ead151d46fd4b