DAFS: Avoid FSYNC operations when fileserver is down for salvage
Salvager code should only perform FSYNC operations when salvaging a
single volume (group), because this is the only type of salvage for
which the fileserver is running and responding to FSYNC requests.  (When
salvaging a partition (or all partitions), the fileserver has been
stopped.) This is implemented by making FSYNC operations conditional on
either of the following equivalent conditions:
- singleVolumeNumber != 0
- salvinfo->useFSYNC == 1

However, there are several places where the salvager may call volume
header routines while the fileserver is down:
- VDeleteVolumeDiskHeader
- VCreateVolumeDiskHeader
- VWriteVolumeDiskHeader

For DAFS only, each of these routines attempts to contact the fileserver
to update the volume group cache (VGC) via FSYNC.  If the fileserver is
down for salvaging, this will cause the salvager to hang while waiting
for a reply from the fileserver FSYNC operation.

Instead, add a new useFSYNC argument to these routines, specifying
whether FSYNC may be used.  Adjust all current callers to specify the
correct value for the new argument:
- salvinfo->useFSYNC for all salvager callers
- 1 for all non-salvager callers

Other non-salvager paths to these volume header routines may safely
assume that the fileserver is always up; therefore, specifying
useFSYNC=1 for these ensures that no functional change should be
incurred for them.

Change-Id: I005baa7193d55a25c1ee1e248d9bfbb4ee9fb702