venus: allow update of cell dbs through fs setcell
Currently, the Cache Manager's kernel-resident list of database server
machines for a given cell can be updated through "fs newcell". If the
cell specified by the -name argument does not exist, a new entry is
created. Although this command can be used to introduce completely new
cells, it is often used to update the list of database server machines
of existing cells. In this scenario, if the name of the entry to be
updated is entered incorrectly, the command in question silently creates
a new entry for the "wrong" cell name. To make things worse, the
unwanted entry will only be removed after rebooting the machine.

To avoid the problem mentioned above, add a new option to "fs setcell"
(-updaddrs) dedicated exclusively to update the list of database server
machines of existing cells. If the cell specified by the -cell argument
does not exist, return an error.

Change-Id: Ifb838adde94310667e231d73b695e57654e4fc00