rx: update_nextCid overflow handling is broken
The overflow handling in update_nextCid() produces a rx_nextCid
value of 0x80000001 which itself is an overflow.   When used
to construct the first call of a new connection the connection
id for the call becomes 0x80000002.

If the same connection id is used for multiple connections from
the same endpoint the accepting rx peer will be very confused.

When authenticated connections are used, the CHALLENGE/RESPONSE
will fail because of a mismatch in the connection's callNumber
array.

All communication from a broken initiator to any rx peer will
fail.

The incorrect overflow calculation was introduced by
39b165cdda941181845022c183fea1c7af7e4356 ("Move epoch and cid
generation into the rx core").

This change corrects the overflow value to become

  1 << RX_CIDSHIFT

Change-Id: If36e3aa581d557cc0f4d2d478f84a6593224c3cc