opr_cache: allow non-self-contained entries
Currently, opr_cache assumes that each entry is self-contained. In other
words, it is not safe to use this library if a given entry (or the
members of this entry) points to another piece of data.

To support non-self-contained entries, this commit extends opr_cache
allowing the user to provide the respective callbacks informing the
library how to properly copy the members of fetched entries and how to
release the memory used by them.

Change-Id: I71b70efc38005febd0f9a9cc55328d59f704b68b