afs: fallback to linked cell if volume isn't available
Currently, the communication from the afs client to the fileserver works
as follows:

1. A connection is created (afs_Conn);
2. Remote procedure call is executed (RXAFS_*);
3. The outcome of this RPC is analyzed (afs_Analyze);
4. Depending on the result returned by step 3, come back to step 1;

If the volume in question is busy, offline, or can’t be reached, step 2
fails and an error is returned to the caller. However, in some
architectures, the same volume is replicated between different cells.
Unfortunately, the current version of the client isn’t smart enough to
redirect requests for unavailable volumes (locally) to the cells holding
an available replica.

To make this possible, this commit extends the -linkedcell feature so
requests can be redirected to the linked cell if our local copy isn't
available.

Change-Id: Ie4fb218c08dcea899cfba3013918d653d5242e7b